Combine all the scoring classes into a single file
