+++ pgmpy/estimators/__init__.py
@@ -1,10 +1,7 @@
 from pgmpy.estimators.base import BaseEstimator, ParameterEstimator, StructureEstimator
 from pgmpy.estimators.MLE import MaximumLikelihoodEstimator
 from pgmpy.estimators.BayesianEstimator import BayesianEstimator
-from pgmpy.estimators.StructureScore import StructureScore
-from pgmpy.estimators.K2Score import K2Score
-from pgmpy.estimators.BDeuScore import BDeuScore
-from pgmpy.estimators.BicScore import BicScore
+from pgmpy.estimators.StructureScore import StructureScore, K2Score, BDeuScore, BicScore
 from pgmpy.estimators.ExhaustiveSearch import ExhaustiveSearch
 from pgmpy.estimators.HillClimbSearch import HillClimbSearch
 from pgmpy.estimators.ConstraintBasedEstimator import ConstraintBasedEstimator
